Move sized int enums into `chia.util.misc`
