+++ tests/core/util/test_streamable.py
@@ -17,7 +17,8 @@ from chia.types.blockchain_format.program import Program
 from chia.types.blockchain_format.sized_bytes import bytes4, bytes32
 from chia.types.full_block import FullBlock
 from chia.types.weight_proof import SubEpochChallengeSegment
-from chia.util.ints import Int8Enum, Int16Enum, Int32Enum, UInt8Enum, UInt16Enum, UInt32Enum, uint8, uint32, uint64
+from chia.util.ints import uint8, uint32, uint64
+from chia.util.misc import Int8Enum, Int16Enum, Int32Enum, UInt8Enum, UInt16Enum, UInt32Enum
 from chia.util.streamable import (
     ConversionError,
     DefinitionError,
